vm-grepmail patch for Emacs 21 Hi,
I'm sending you a tiny patch I had to use in order to get vm-grepmail
to work with a recent CVS version of GNU Emacs 21 (and probably the
released versions as well).

It does two things:

i) Defines vm-grepmail-folders-history to be a function to stop
read-file-name throwing an error. VM does the same thing with
vm-folder-history.

ii) I found M-x vm-grepmail never stopped prompting me for folders.
This was because I have vm-folder-directory = "/home/gmorris/VM", but
the value returned by read-file-name is abbreviated to "~/VM". As a
result, the `(while (not (string= fd default))' loop never exited.

Cheers,

Glenn