Bugfix: vm-shrunken-headers was causing folder corruption.
For messages which were directly displayed from the folder buffer it was
causing "*"s to be inserted into the folder buffer at the position of the
visibility widget.

The fix is to ensure we are only working on presentation copies of the
message.

Thanks to RPG for reporting and hunting this bug, since I would have
never noticed it as with activated filling I always get presentation
copies, also for plain text messages.