Update occurrences in `post-command-hook'
Before the occurrences are updated in `modification-hooks' directly.  Some
commands adjust the current point after `modification-hooks' based on the saved
position, e.g. yank, replace-match. Then the point might end up to a wrong
place.

In this commit, a hook is added into `post-command-hook' when iedit-mode is
acitve.  The change in `modification-hooks' are collected into a
`iedit-after-change-list'.  In the hook, the list are accumlated into one change
and then applied to all the other occurrences.