Change the way the current completion is highlighted
This change should eliminate any differences between the highlighted
completion and the actual completion string.

* vcomplete.el (vcomplete--highlight-completion-at-point): New
function which highlights the completion at point using the same
method used by 'choose-completion' to get the bounds of the completion
string.
(vcomplete--move-n-completions): Use the new function instead of the
previous method of highlighting the completion.