Highlight the completion at point in the completion list buffer
This is different from previous behavior since the completion at point
is highlighted even when navigated to manually.
Note that this doesn't work when completing in-region, although this
doesn't really matter since switching buffers exits
'completion-in-region-mode'.

* vcomplete.el (vcomplete-next-completion, vcomplete-prev-completion):
Don't set 'this-command' to 'vcomplete--no-update'.
(vcomplete--last-string): New variable to hold the last pending
completion string.
(vcomplete--last-string-in-minibuffer): New function which sets
'vcomplete--last-string' (when completing in a minibuffer).
(vcomplete--last-string-in-region): New function which sets
'vcomplete--last-string' (when completing in-region).
(vcomplete--update-in-minibuffer, vcomplete--update-in-region): Modify
to update when the current completion string is different to
'vcomplete--last-string'.
(vcomplete--reset-vars): Add steps to reset 'vcomplete--last-string'
and remove 'vcomplete--highlight-completion-at-point' from the local
'post-command-hook'.
(vcomplete--setup-completions): New function which sets up the
completion list buffer for highlighting at point.
(vcomplete--setup-current): Rename from 'vcomplete--setup'.
Restructure and add steps to add the "last-string" functions added to
the local 'pre-command-hook'.
(vcomplete-mode): Rename references to 'vcomplete--setup' and
add steps to add/remove 'vcomplete--setup-completions' to
'completions-list-mode-hook'.