FIX local mentions/links checking
we remove the hacky non-check for a user handle from mastodon-tl--process-link, so that it
/always/ runs mastodon-tl--extract-userhandle-from-url.

the in extract-userhandle-from-url, we test if the url host = local instance.
if so, we just return buffer-text, which = "@user", with no suffix. else we
return a full "@user@instance.url" handle.

remove unused var after fix to userhandle parsing