re-implement user auth and token generation mechanism
Refactor `mastodon-auth--generate-token' to work with new user
authentication and token generation mechanism.  This enables mastodon.el
to work with 2FA and also not to handle password directly.

In this implementation mastodon-auth--generate-token gets authorization
code from the user and sends post request to mastodon server.  Ask for
authorization code from the user using two helper functions:

* mastodon-auth--ask-authorization-code: Explain to the user what the
user needs to do to get the authorization code.  Store this explanation
message in variable `mastodon-auth--explanation'.

* mastodon-auth--get-browser-login-url: Return a appropriate query
string appended to url to the caller, which is needed by the user to
access the authorization code.