Remove dependency on `f` and fix some compilation issues
* test/test-helper.el: Use lexical-binding.  Mark as not to be compiled.

* test/pacmacs-walls-test.el: Use lexical-binding.  Require `el-mock`,
`dash`, and `pacmacs-walls`.  Mark as not to be compiled.

* test/pacmacs-vector-test.el: Use lexical-binding.
Require `pacmacs-vector`.

* test/pacmacs-utils-test.el: Use lexical-binding.  Require `pacmacs-utils`.

* test/pacmacs-test.el: Use lexical-binding.  Require `el-mock` and
`pacmacs`.  Mark as not to be compiled.

* test/pacmacs-score-test.el: Use lexical-binding.
Require `pacmacs-score`.

* test/pacmacs-render-test.el: Use lexical-binding.  Require `el-mock`.
Mark as not to be compiled.

* test/pacmacs-board-test.el: Use lexical-binding.  Require `pacmacs-board`.

* test/pacmacs-anim-test.el: Use lexical-binding.  Require `el-mock`.
Mark as not to be compiled.

* pacmacs.el: Remove `f` and `cl-lib` from Package-Requires.
`cl-lib` was redundant since it's included in Emacs≥24.4.
Don't require `f` any more.
(pacmacs-levels-folder): Remove redundant `:group`.
(pacmacs-mode-map, pacmacs-game-over-mode-map): Define it once and for
all so we don't overwrite user changes when the mode is called again.
(pacmacs-scores-mode): Don't change the `q` binding in
`pacmacs-game-over-mode-map`.
(pacmacs-quit): Delete scrobe buffer if applicable, instead.

* pacmacs-vector.el: Require `dash` to fix compilation errors.

* pacmacs-score.el: Don't require `f`.
(pacmacs--read-score-table): Read directly from a buffer rather than
going through a string.

* tools/pacmacs-rr.el: Don't require `f`.
(pacmacs--load-test-case): Read directly from a buffer rather than
going through a string.
(pacmacs-it-recorder-mode-map): Define it once and for all so we don't
overwrite user changes when the mode is called again.

* tools/compile.el: Use lexical-binding and don't require `f`.
Require `dash` at top-level so it's byte-compiled properly.
Don't require `cask` during compilation of the file itself.

* tools/att.el: Use lexical-binding and don't require `f`.
(att-replayer-finished): Use `write-region` instead of `f-write`.

* .gitignore: Add ELPA's generated files.
* it-cases/it-case0*.el: Change extentions from `.el` to `.eld` to
indicate that those files do not contain ELisp code but ELisp data
(they should use `lisp-data-mode` rather than `emacs-lisp-mode` and
they should not be byte-compiled).
* tools/att.el (att-it-case-path): Adjust accordingly.