improve propertizing of boost/fave markers
- previously, boost/fave markers behaved differently depending on whether they
had just been created by the user, or if they were already existing and loaded
with the timeline. in the first case, they'd not be part of the byline, in the
second they would be.

we make it that they not part of the byline, so `mastodon-tl--goto-next-toot',
which works according to text properties, should always put point after them,
on the author-byline.

this also means that we can add help-echos and actions to the author byline
without having to worry about also adding them to the boost/fave markers.

fix call to format-faved-or-boosted-byline