test for toot-json before following/muting/blocking etc
this is more needed now that we have some other buffers like follow
suggestions and follow requests, which want to be able to use these functions.

previously, if any of them were called in these buffers when not on the
correct item, the buffer would reload and duplicate.

ultimately this was probably due to tl--property calls failing, which causes
goto-next-toot to be called, which when it also fails calls tl--more.