auth: present auth code request in a nicer way
Instead of using the minibuffer prompt to present the explanation what
the user needs to do to get authorization code, use Emacs buffer and
Window.  The minibuffer prompt should have 2-3 words.

This increases the readability of the explanation.

Of course delete explanation buffer and restore the window configuration
as it was after reading from minibuffer.