auth: make old mastodon.el users aware
Most old users of the mastodon.el wouldn't know about the introduction
of new variable `mastodon-active-user'[1].  Our goal is to make them
aware.

Set the default value of [1] to nil.  This way we can know that the user
hasn't set [1] properly because user is unaware of it.

In the definition of `mastodon-auth--access-token', check [1]'s value
and if it's null, explain to the user about the new situation. Finally
signal the error "Variables not set properly".

We have considered the possibility of automating the process but since
the new login mechanism depends on setting [1] *in the init file*, the
only way to automate it would be to write to user's init file which we
consider to be very rude and shouldn't be done by this package.