make timestamps for boost/fave notifs the time of orig toot
do not use the time stamp of the boosting/faving.

in other views, mastodon-tl--field already handles this by fetching from
'reblog if present, but in notifs, there's no reblog section, and instead
there's a status section. so we fetch from status if present, else fallback to
using --field.

this became necessary when we started attaching the boost/fave json to such
notifs, rather than the json of the boosed/faved toot.