paginate bookmarks timeline (3/4 working?)
seems tl--more seems to repeat bookmarks after 3-4 loads. a similar issue was
solved with favourites, but it was always after 2nd load. and we are using the
fixed code here already. i have many more favourites than bookmarks. but still
it would be nice to not seem to load more when we hit the end.