mastodon-tl--spoiler: Add error handling to /preferences api query
mastodon.el currently breaks when trying to load a spoilered toot if
the /preferences endpoint returns something unexpected, preventing
mastodon-tl--buffer-spec from being set and not automatically enabling
the major mode. This is particularly a problem on GoToSocial, which
does not currently have the endpoint implemented and instead returns a
404 HTML page, causing a JSON parsing error.

This PR adds a simple check which causes the API check in
mastodon-tl--spoiler to return null if it errors, thus triggering the
default behavior of hiding CW'd toots until opened.