Don't expose the `-*-functions` vars in the macroexpanded code
Set those vars from the `mock--stub-setup` and `mock/setup` functions
instead.  This has the benefit of simplifying the macroexpanded code
a bit, but it also causes the `mock` and `not-called` functions
to be added to `-stubbed-functions` which was not the case any more.
This has the advantage that we don't need to call
(mapc #'mock/teardown -mocked-functions) explicitly any more.
It has the subtle consequence of removing the mocks and not-called
*before* calling `mock-verify` rather than afterwards rather than
afterwards, but AFAICT, `mock-verify` does not need those mocks/not-called
to be still active, so it seems to be a safe change.

* el-mock.el (mock--stub-setup): Add func to `-stubbed-functions`.
(stub): Don't add func to `-stubbed-functions` any more
since it's done by `mock--stub-setup` now.
(mock/setup): Add func to `-mocked-functions`.
(mock): Don't add func to `-mocked-functions` any more since it's done
by `mock/setup` now.
(not-called): Don't add func to `-mocked-functions` any more since it
was useless anyway.
(mock/teardown): Delete (now) unused function.