tl--prev-toot-id: don't assume our funs return anything.
fixes a bug that would break thread loading, perhaps because an ancestor toot
wasn't there when it was supposed to be, so we were check for prev id when
there was no prev toot. maybe caused by viewing threads that include toots by
ppl on blocked servers, which don't display.