Tests now indicate where they are loading totp-auth et al from
‘require’ won't reload a feature that's already been loaded
so if you test totp-auth in a copy of emacs that already has
totp-auth loaded from elsewhere (eg a stable version instead
of the source repo) then that's what will get tested.

This may or may not be desireable so we log which copy we loaded
and leave it up to the caller to decide if that's good or bad.