(editorconfig-indentation-alist): Make it more declarative
Make the functions in `editorconfig-indentation-alist` return the list
of (VAR . VAL) to set rather than letting them set the vars directly.
This makes it possible to obey `editorconfig--should-set` for those
vars, and more importantly makes it usable within for Emacs's
hack-dir-local system where we manipulate such lists of settings
before we apply them.